Moved knex db instance
