+++ server/src/Database/index.ts
@@ -1,5 +1,5 @@
 import knex from 'knex';
-import config from './Database/knexfile.js';
+import config from './knexfile.js';
 
 const db = knex(config.development);
